Use mongodb message queue to work with Neso
I may be overusing a single data storage for multiple pro-
blems, either way mongoDB offers a good enough infrastruc-
ture to use it as a message queue.

Despite the fact the initial intention of Sao was to handle
the Entry handling, transform and processing with CMS
through its UI (meaning to have resend user authentication
tokens and parsing html), I have decided to delegate these
responsabilities to another component (cms-neso) in favor
of maintanibility and ease of implementation (parsing html
is in fact a trivial task in a JS application).