Groom /submit-* action endpoints
After a entry/draft submit process request is successfully
received it used to reply a full transaction object alongside a
201 status and a new location href, with the purpose of letting
the client to follow and poll from it. In other words, the 99% of
the time the client will ignore the initial reply, based on that
I chose to reduce the amount shown by it.